Move Tokenier classes to separted files
